fix: Change commands distribution Closes #13
