+++ discordbot/src/commands/token_commands.py
@@ -1,18 +1,9 @@
 from discordbot.src.setup import client
 import discord
-from . import Session
+from .. import Session
 from src.database.models import UserToken
 
-@client.event
-async def on_ready():
-    print(f'Discord Bot {client.user} activated')
-    
-    
-@client.command(help="Say hello to check if the bot is working correctly")
-async def hello(ctx):
-    await ctx.send("Hello!")
-    
-    
+
 @client.command(help="Configurate your access token. This command makes the bot send you a private message to configurate your token")
 async def token_config(ctx):
     user = ctx.message.author
